refactor(github): useGithubForm to a hook that returns getter + setter
