+++ utils/formCache.ts
@@ -1,16 +1,14 @@
-export const getCachedFormData = (fileRelativePath: string) => {
+export const getCachedFormData = (id: string) => {
   if (typeof localStorage === 'undefined') {
     return {}
   }
-  return JSON.parse(localStorage.getItem(fileRelativePath) || '{}')
+  return JSON.parse(localStorage.getItem(id) || '{}')
 }
 
-export const setCachedFormData = (
-  fileRelativePath: string,
-  data: { sha: string }
-) => {
+export const setCachedFormData = (id: string, data: { sha: string }) => {
+  console.log(data)
   if (typeof localStorage === 'undefined') {
     return
   }
-  localStorage.setItem(fileRelativePath, JSON.stringify(data))
+  localStorage.setItem(id, JSON.stringify(data))
 }
