+++ utils/github/useLocalGithubJsonForm.ts
@@ -1,6 +1,5 @@
 import { FormOptions, useLocalForm, usePlugins, Field, useCMS } from 'tinacms'
-import { getCachedFormData, setCachedFormData } from '../formCache'
-import { useGithubForm, GithubOptions, GitFile } from './useGithubForm'
+import { useGitFileSha, GithubOptions, GitFile } from './useGitFileSha'
 import { FORM_ERROR } from 'final-form'
 import OpenAuthoringError from '../../open-authoring/OpenAuthoringError'
 import { getForkName } from '../../open-authoring/open-authoring/repository'
@@ -18,7 +17,7 @@ const useGithubJsonForm = <T = any>(
   githubOptions: GithubOptions
 ) => {
   const cms = useCMS()
-  useGithubForm(jsonFile)
+  const [getSha, setSha] = useGitFileSha(jsonFile)
 
   const [formData, form] = useLocalForm({
     id: jsonFile.fileRelativePath, // needs to be unique
@@ -32,7 +31,7 @@ const useGithubJsonForm = <T = any>(
           githubOptions.forkFullName,
           githubOptions.branch,
           jsonFile.fileRelativePath,
-          getCachedFormData(jsonFile.fileRelativePath).sha,
+          getSha(),
           JSON.stringify(formData, null, 2),
           'Update from TinaCMS'
         )
@@ -41,9 +40,7 @@ const useGithubJsonForm = <T = any>(
             `Saved Successfully: Changes committed to ${getForkName()}`
           )
 
-          setCachedFormData(jsonFile.fileRelativePath, {
-            sha: response.content.sha,
-          })
+          setSha(response.content.sha)
         })
         .catch(e => {
           return { [FORM_ERROR]: new OpenAuthoringError(e.message, e.status) }
